repo tests conversion to pytest complete, replaced original testfile
